+++ controler/AdminControler.php
@@ -58,7 +58,7 @@ class AdminControler extends Controler
         $this->getRoleSQL()->edit('admin', 'Administrateur');
 
         foreach ($this->getRoleDroit()->getAllDroit() as $droit) {
-            if (!str_contains($droit, '-destinataire') && !str_contains($droit, '-response')) {
+            if (!str_contains($droit, '-destinataire:') && !str_contains($droit, '-reponse:')) {
                 $this->getRoleSQL()->addDroit('admin', $droit);
             }
         }